+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # REPLACE-GLOBAL-WITH-NEXT-VERSION
 - MIG-100 - Add migration profile for Shopware 5.7
+- MIG-243 - Fix migration of orders for Shopware 6 profile
 - MIG-247 - Fix write protection errors for Shopware 6 migration
 
 # 4.1.1